intellisense service moved to intellisense namespace
