+++ BeaverSoft.Texo.Core/TexoEngineBuilder.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using BeaverSoft.Texo.Core.Actions;
 using BeaverSoft.Texo.Core.Actions.Implementations;
 using BeaverSoft.Texo.Core.Commands;
@@ -7,6 +7,7 @@ using BeaverSoft.Texo.Core.Environment;
 using BeaverSoft.Texo.Core.Help;
 using BeaverSoft.Texo.Core.Input;
 using BeaverSoft.Texo.Core.Input.History;
+using BeaverSoft.Texo.Core.Intellisense;
 using BeaverSoft.Texo.Core.Logging;
 using BeaverSoft.Texo.Core.Runtime;
 using BeaverSoft.Texo.Core.View;
